+++ src/router/__init__.py
@@ -1,0 +1,1 @@
+"""Action router package."""